Fixed some minor mistakes

+++ X10D.Performant/src/StringExtensions/System.Byte.cs
@@ -8,7 +8,7 @@ namespace X10D.Performant
     public static partial class StringExtensions
     {
         /// <inheritdoc cref="byte.Parse(string,NumberStyles,IFormatProvider)"/>
-        public static short ToByte(this string value, NumberStyles style = NumberStyles.Integer, IFormatProvider? formatProvider = null) =>
+        public static byte ToByte(this string value, NumberStyles style = NumberStyles.Integer, IFormatProvider? formatProvider = null) =>
             byte.Parse(value, style, formatProvider ?? NumberFormatInfo.CurrentInfo);
 
         /// <inheritdoc cref="byte.TryParse(string,NumberStyles,IFormatProvider,out byte)"/>
